legacy gcp cloud function/update-labels/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.25.*
pandas==2.2.*
db-dtypes==1.3.*
google-api-python-client==2.151.*
numpy==2.1.*
google-cloud-bigquery-storage==2.27.*
protobuf==5.28.*
pandas-gbq==0.24.*
flask==3.1.*